+++ src/components/molecules/unique/smartSpinner/smartSpinnerConsts.ts
@@ -1,0 +1,2 @@
+export const QUANTITY_UNKNOWN = null;
+export const TIME_UNKNOWN = null;
