+++ src/components/molecules/unique/smartSpinner/smartSpinnerEnums.ts
@@ -14,6 +14,3 @@ export enum SpinnerTextPosition {
     OverSpinner = 1,
     BelowSpinner = 2
 }
-
-export const QUANTITY_UNKNOWN = null;
-export const TIME_UNKNOWN = null;
